Created `Function.prototype.once.js` which contains a function to generate functions that can only be called `n` times.
When `onceFn = fn.once(n)` is called, it returns a function
that will only execute the first `n` times it's called using
the parameters provided to `onceFn`. If no parameter is
provided, `n = 1`.

If `onceFn` can still be called, the return value of `onceFn`
is the return value of `fn`, otherwise the second, optional,
`ret` parameter is used as the return value.