Created `Function.prototype.throttle.js` to generate functions that are throttled to prevent being called more than once in the specified period.
When `throttledFn = fn.throttle(t, ret)` is called, it returns a
function that will be executed with the parameters provided only
if it has not been called less than `t` milliseconds ago.

The return value of `throttledFn` is the return value of `fn`
unless `t` milliseconds haven't elapsed sinse the last call to `fn`
in which case the return value is the optional `ret` parameter.