Created `String.prototype.tmpl.js` to use on string templates as a simple string replacement mechanism for variable data.
When `str = template.tmpl(data)` is called, the `template` string
will have all occurances of `<%=prop%>` replaced with `data.prop`
where prop is any unquoted string.

Additionally, all occurances of `<%="string"%> will be replaced by
the value contained within the `"` characters. This is primarily to
be used to provide literal `<%=` and `%>` tags in the output.